requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
lxml>=4.9.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
jinja2>=3.1.0
schedule>=1.2.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
